streamlit/requirements.txt
Streamlit
Pandas
Boto3
Langchain 
langchain-core
langchain-openai
langchain-qdrant
python-dotenv
